backend/requirements.dev.txt
django==4.2
djangorestframework==3.14.0
pytz==2022.1
psycopg2-binary==2.9.3
django-cors-headers